refactor: Change Layout dir and component name
